+++ tests/compile-fail/ClrEnumNonEnumClass.kt
@@ -1,0 +1,13 @@
+import kotlin.clr.ClrEnum
+
+@ClrEnum
+class ClrEnumNonEnumClass
+
+@ClrEnum
+interface ClrEnumNonEnumInterface
+
+@ClrEnum
+object ClrEnumNonEnumObject
+
+@ClrEnum
+annotation class ClrEnumNonEnumAnnotation